requirements.txt
streamlit>=1.37,<2.0
paddleocr>=2.7.0.3
openai>=1.45.0
langchain>=0.2.10
langchain-openai>=0.1.7
pandas>=2.0
plotly>=5.18
pillow>=10.0
python-dotenv>=1.0
pydantic>=2.0
numpy>=1.26
scikit-learn>=1.4
